# Use intervals for idle timeouts