This PR should be merged only after #6.

It contains suggestion from #6 and mitigate the problem of using non-monotonic clock usage, that it was before.

Benchmark before (from #6):
```
     Running benches/bench.rs (target/release/deps/bench-a5e9977bff1da307)

running 2 tests
test tests::bench_mux_send ... bench:      32,561.84 ns/iter (+/- 6,245.88) = 4025 MB/s
test tests::bench_tcp_send ... bench:      12,072.64 ns/iter (+/- 901.40) = 10857 MB/s

test result: ok. 0 passed; 0 failed; 0 ignored; 2 measured; 0 filtered out; finished in 67.56s
```

Benchmark after:
```
     Running benches/bench.rs (target/release/deps/bench-a5e9977bff1da307)

running 2 tests
test tests::bench_mux_send ... bench:      32,121.46 ns/iter (+/- 5,845.00) = 4080 MB/s
test tests::bench_tcp_send ... bench:      11,974.32 ns/iter (+/- 854.97) = 10946 MB/s

test result: ok. 0 passed; 0 failed; 0 ignored; 2 measured; 0 filtered out; finished in 71.84s
```